insurance_data_classifier/requirements.txt
# Core ML
scikit-learn>=1.2.0
torch>=2.0.0
transformers>=4.30.0

# Data processing
numpy>=1.21.0
pandas>=1.5.0

# Configuration
pyyaml>=6.0

# Hugging Face utilities
accelerate>=0.20.0
datasets>=2.10.0
protobuf>=3.20.0
tokenizers>=0.13.0

# Utilities
tqdm>=4.64.0